Move funcs_deriveODEsFromRules, _deriveMasterEquation and _doVanKampenExpansion from models.py to views.py
